massive restructure for shadcn and file naming
<!-- ps-id: 4432b371-edeb-45e2-89b1-fb75b98a61c9 -->

turbo upgrade

<!-- ps-id: ca16f290-09c0-4565-8c0a-daf7e1a1a828 -->
