+++ apps/admin-client/src/auth/password-reset-page.tsx
@@ -1,7 +1,7 @@
 import { HttpMethod } from '@/common/custom-fetcher';
 import { LOGIN } from '@/common/routes';
 import { Button } from '@/components/button';
-import { Input } from '@/components/Input';
+import { Input } from '@/components/input';
 import { Label } from '@/components/label';
 import { useToast } from '@/components/toast/use-toast';
 import { useMutation } from '@tanstack/react-query';
