+++ apps/admin-client/src/components/fields/text-field.tsx
@@ -1,23 +1,22 @@
 import { ChangeEvent } from 'react';
 import { AdminModelField } from '@repo/types';
-import FieldLabel from '@/components/record-field-label';
-import { Input } from '@/components/Input';
+import FieldLabel from '@/features/records/record-field-label';
+import { Textarea } from '@/components/textarea';
 
 interface Props {
   field: AdminModelField;
   value: string;
   onChange: (key: string, value: string) => void;
 }
-export default function StringField({ field, value, onChange }: Props) {
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+export default function TextField({ field, value, onChange }: Props) {
+  const handleChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     onChange(field.name, e.currentTarget.value);
   };
 
   return (
     <div>
       <FieldLabel field={field} />
-      <Input
-        type="text"
+      <Textarea
         id={field.name}
         name={field.name}
         value={value || ''}
