+++ apps/admin-client/src/features/records/delete-record-button.tsx
