+++ apps/admin-client/src/features/records/show-view-field.tsx
