+++ apps/admin-client/src/layout/root-layout.tsx
@@ -9,7 +9,7 @@ import {
 } from '@heroicons/react/24/outline';
 import { classNames } from '@repo/utils';
 import { HOME, LOGOUT, PROFILE } from '@/common/routes';
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/Avatar';
+import { Avatar, AvatarFallback, AvatarImage } from '@/components/avatar';
 import { AuthProvider } from '@repo/auth';
 
 type NavItem = {
