+++ apps/admin-client/src/main.tsx
@@ -1,6 +1,6 @@
-import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
-import "./index.css";
+import ReactDOM from 'react-dom/client';
+import App from './app.tsx';
+import './index.css';
 
-const el = document.getElementById("root")!;
+const el = document.getElementById('root')!;
 ReactDOM.createRoot(el).render(<App />);
