+++ apps/client/src/auth/forgot-password-page.tsx
@@ -1,6 +1,8 @@
 import { HttpMethod } from '@/common/custom-fetcher';
 import { LOGIN } from '@/common/routes';
 import { Button } from '@/components/button';
+import { Input } from '@/components/input';
+import { Label } from '@/components/label';
 import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
 import { Form, Link } from 'react-router-dom';
@@ -56,18 +58,12 @@ export default function ForgotPasswordPage() {
         <div className="bg-white px-6 py-12 shadow sm:rounded-lg sm:px-12">
           <Form className="space-y-6" onSubmit={submit}>
             <div>
-              <label
-                htmlFor="email"
-                className="block text-sm font-medium leading-6 text-gray-900"
-              >
-                Email
-              </label>
+              <Label htmlFor="email">Email</Label>
               <div className="mt-2">
-                <input
+                <Input
                   name="email"
                   type="email"
                   autoComplete="email"
-                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
                   required
                   value={email}
                   onChange={(e) => setEmail(e.currentTarget.value)}
