+++ apps/client/src/auth/login-page.tsx
@@ -1,5 +1,7 @@
 import { FORGOT_PASSWORD } from '@/common/routes';
 import { Button } from '@/components/button';
+import { Input } from '@/components/input';
+import { Label } from '@/components/label';
 import { useState } from 'react';
 import {
   useLocation,
@@ -42,18 +44,13 @@ function Login() {
             <input type="hidden" name="redirectTo" value={from} />
 
             <div>
-              <label
-                htmlFor="email"
-                className="block text-sm font-medium leading-6 text-gray-900"
-              >
-                Email
-              </label>
+              <Label htmlFor="email">Email</Label>
               <div className="mt-2">
-                <input
+                <Input
+                  id="email"
                   name="email"
                   type="email"
                   autoComplete="email"
-                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
                   required
                   value={data.email}
                   onChange={(e) =>
@@ -67,19 +64,13 @@ function Login() {
             </div>
 
             <div>
-              <label
-                htmlFor="password"
-                className="block text-sm font-medium leading-6 text-gray-900"
-              >
-                Password
-              </label>
+              <Label htmlFor="password">Password</Label>
               <div className="mt-2">
-                <input
+                <Input
                   id="password"
                   name="password"
                   type="password"
                   autoComplete="current-password"
-                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
                   required
                   value={data.password}
                   onChange={(e) =>
