+++ apps/client/src/auth/protected-loader.ts
