+++ package.json
@@ -58,6 +58,6 @@
   "devDependencies": {
     "prettier": "^3.1.1",
     "prettier-plugin-tailwindcss": "^0.5.11",
-    "turbo": "^1.12.5"
+    "turbo": "^1.13.2"
   }
 }
